add Params module
