+++ lib/snowflex/worker.ex
@@ -5,17 +5,11 @@ defmodule Snowflex.Worker do
 
   use GenServer
 
+  alias Snowflex.Params
   alias Snowflex.Telemetry
 
   @timeout :timer.seconds(60)
   @gc_delay_ms 5
-  @string_types ~w(
-    sql_char
-    sql_wchar
-    sql_varchar
-    sql_wvarchar
-    sql_wlongvarchar
-  )a
 
   def start_link(connection_args) do
     GenServer.start_link(__MODULE__, connection_args, [])
@@ -142,7 +136,7 @@ defmodule Snowflex.Worker do
 
   defp do_param_query(%{pid: pid} = state, query, params) do
     ch_query = to_charlist(query)
-    ch_params = prepare_params(params)
+    ch_params = Params.prepare(params)
 
     case :odbc.param_query(pid, ch_query, ch_params) do
       {:error, reason} ->
@@ -152,38 +146,6 @@ defmodule Snowflex.Worker do
       result ->
         {{:ok, result}, state}
     end
-  end
-
-  defp prepare_params(params) do
-    Enum.map(params, &prepare_param/1)
-  end
-
-  defp prepare_param({type, values}) when not is_list(values) do
-    prepare_param({type, [values]})
-  end
-
-  defp prepare_param({{type_atom, _size} = type, values}) when type_atom in @string_types do
-    {type, Enum.map(values, &null_or_charlist/1)}
-  end
-
-  defp prepare_param({type, values}) do
-    {type, Enum.map(values, &null_or_any/1)}
-  end
-
-  defp null_or_charlist(nil) do
-    :null
-  end
-
-  defp null_or_charlist(val) do
-    to_charlist(val)
-  end
-
-  defp null_or_any(nil) do
-    :null
-  end
-
-  defp null_or_any(any) do
-    any
   end
 
   defp send_heartbeat(state) do
